refactor: remove unnedeed controller
